# openapi.yml
openapi: "3.0.1"
info:
  description: "Servidor de l'aplicació fithaus de l'assignatura PES. Aquí podràs trobar operacions per gestionar usuaris del sistema."
  version: "1.0.0"
  title: "SERVIDOR FITHAUS"
  contact:
    email: "fithaus2021@gmail.com"
servers:
  - url: "127.0.0.1:8000"
paths:
...
  /user/{id}:
      get:
        tags:
          - "user"
        summary: "Obtenir l'usuari amb l'id especificat"
        description: "Proporcionar l'id de l'usuari"
        operationId: "getUserById"
        parameters:
          - name: "id"
            in: "path"
            description: "Id de l'usuari"
            required: true
            schema:
              type: "string"
        responses:
          "200":
            description: "Operació exitosa"
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/User'
          "400":
            description: "Alguna cosa no ha anat bé..."
      patch:
        tags:
          - "user"
        summary: "Actualitzar info d'un usuari"
        description: "Especificar només els camps a actualitzar"
        operationId: "updateUser"
        requestBody:
          description: "Proporcionar els camps a actualitzar"
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: "string"
                    maxLength: 200
                  firstName:
                    type: "string"
                    maxLength: 200
                  lastName:
                    type: "string"
                    maxLength: 200
                  email:
                    type: "string"
                    maxLength: 200
                    format: "email"
                  password:
                    type: "string"
                    maxLength: 200
                    format: "password"
                  gender:
                    type: "string"
                    enum: [ M, W, X ]
                  birthdate:
                    type: "string"
                    maxLength: 200
                    format: "date"
                  activitiesdone:
                    type: "integer"
                    default: 0
                  achievements:
                    type: "string"
                    maxLength: 200
                  points:
                    type: "integer"
                    default: 0
                  level:
                    type: "integer"
                    default: 0
                  objective:
                    type: "string"
                    maxLength: 200
                  interestcategories:
                    type: "string"
                    maxLength: 200
                  weight:
                    type: "integer"
                    default: 0
                  height:
                    type: "integer"
                    default: 0
                  imc:
                    type: "integer"
                    default: 0
                  igc:
                    type: "integer"
                    default: 0
        responses:
          "200":
            description: "Operació exitosa"
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/User'
          "400":
            description: "Usuari no trobat"
      delete:
        tags:
          - "user"
        summary: "Eliminar usuari"
        description: "Proporcionar l'id de l'usuari"
        operationId: "deleteUser"
        parameters:
          - name: "id"
            in: "path"
            description: "Id de l'usuari a eliminar"
            required: true
            schema:
              type: "string"
        responses:
          "200":
            description: "Operació exitosa"
          "404":
            description: "Usuari no trobat"
...
components:
  schemas:
    User:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        username:
          type: "string"
          maxLength: 200
        firstName:
          type: "string"
          maxLength: 200
        lastName:
          type: "string"
          maxLength: 200
        email:
          type: "string"
          maxLength: 200
          format: "email"
        password:
          type: "string"
          maxLength: 200
          format: "password"
        gender:
          type: "string"
          enum: [M, W, X]
        birthdate:
          type: "string"
          maxLength: 200
          format: "date"
        activitiesdone:
          type: "integer"
          default: 0
        achievements:
          type: "string"
          maxLength: 200
        points:
          type: "integer"
          default: 0
        level:
          type: "string"
          enum: [B, I, A]
        objective:
          type: "object"
        interestcategories:
          type: "object"
        weight:
          type: "number"
          default: 0.0
        height:
          type: "number"
          default: 0.0
        imc:
          type: "number"
          default: 0.0
        igc:
          type: "number"
          default: 0.0
...